third_party/blink/web_tests/platform/generic/fast/parser/attributes-on-close-script-expected.txt

ALERT: PASS (as long as this test doesn't trigger an ASSERT)
| <html>
|   <head>
|     <script>
|       src="../../resources/dump-as-markup.js"
|     "
"
|     <script>
|       "
alert("PASS (as long as this test doesn't trigger an ASSERT)");
"
|     "
"
|   <body>
